keypair utility 置き換え
